Add Mapstruct mapping and validation exception handling

+++ user-service/src/main/java/com/prathmesh/spendwise/userservice/service/impl/UserServiceImpl.java
@@ -3,6 +3,8 @@ package com.prathmesh.spendwise.userservice.service.impl;
 import com.prathmesh.spendwise.userservice.dto.request.UserRequest;
 import com.prathmesh.spendwise.userservice.dto.response.UserResponse;
 import com.prathmesh.spendwise.userservice.entity.User;
+import com.prathmesh.spendwise.userservice.exception.ResourceNotFoundException;
+import com.prathmesh.spendwise.userservice.mapper.UserMapper;
 import com.prathmesh.spendwise.userservice.repository.UserRepository;
 import com.prathmesh.spendwise.userservice.service.UserService;
 import org.springframework.stereotype.Service;
@@ -16,23 +18,20 @@ import java.util.Map;
 public class UserServiceImpl implements UserService {
 
     private final UserRepository userRepository;
+    private final UserMapper userMapper;
 
     private final Map<Long, User> users = new HashMap<>();
     private Long idCounter = 1L;
 
-    public UserServiceImpl(UserRepository userRepository) {
+    public UserServiceImpl(UserRepository userRepository, UserMapper userMapper) {
         this.userRepository = userRepository;
+        this.userMapper = userMapper;
     }
 
     @Override
     public UserResponse createUser(UserRequest request) {
 
-        User user = new User();
-
-        user.setFirstName(request.getFirstName());
-        user.setLastName(request.getLastName());
-        user.setEmail(request.getEmail());
-        user.setPhone(request.getPhone());
+        User user = userMapper.toEntity(request);
 
         User savedUser = userRepository.save(user);
 
@@ -52,7 +51,8 @@ public class UserServiceImpl implements UserService {
     public UserResponse getUserById(Long id) {
 
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User not found with Id : "+id));
 
         return mapToResponse(user);
     }
@@ -61,7 +61,9 @@ public class UserServiceImpl implements UserService {
     public UserResponse updateUser(Long id, UserRequest request) {
 
         User user = userRepository.findById(id)
-                .orElseThrow(() -> new RuntimeException("User not found"));
+                .orElseThrow(() ->
+                        new ResourceNotFoundException("User not found with Id : "+id)
+                );
 
         user.setFirstName(request.getFirstName());
         user.setLastName(request.getLastName());
@@ -76,17 +78,15 @@ public class UserServiceImpl implements UserService {
     @Override
     public void deleteUser(Long id) {
 
+        if (!userRepository.existsById(id)){
+            throw new ResourceNotFoundException("User not found with Id : "+id);
+        }
+
         userRepository.deleteById(id);
     }
 
     private UserResponse mapToResponse(User user) {
 
-        return new UserResponse(
-                user.getId(),
-                user.getFirstName(),
-                user.getLastName(),
-                user.getEmail(),
-                user.getPhone()
-        );
+        return userMapper.toResponse(user);
     }
 }
